refactor: align package structure and console output with Laravel conventions
